third_party/blink/web_tests/platform/generic/fast/url/safari-extension-expected.txt

Verify the canonicalization of Safari Extension URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS successfullyParsed is true

TEST COMPLETE
